Colour codes

+++ post-install.sh
@@ -1,6 +1,10 @@
 #!/bin/bash
 set -e
 
+RED='\033[0;31m'
+YEL='\033[1;33m'
+NC='\033[0m' # No Color
+
 # As root
 # adduser yellow sudo
 
@@ -15,7 +19,7 @@ set -e
 
 if [ ! -f ~/.gnupg/gpg.conf ]
 then
-  echo "Creating GPG key."
+  echo -e "${YEL}Creating GPG key.${NC}"
   mkdir -p ~/.gnupg/
   cat >> ~/.gnupg/gpg.conf <<EOF
 personal-digest-preferences SHA256
@@ -27,13 +31,13 @@ fi
 
 if ! grep "contrib non-free" /etc/apt/sources.list;
 then
-  echo "Adding contrib and non-free sources for nvidia drivers."
+  echo -e "${YEL}Adding contrib and non-free sources for nvidia drivers.${NC}"
   sudo sed --in-place=.old -e 's/main$/main contrib non-free/g' /etc/apt/sources.list
 fi
 sudo apt-get update
 sudo apt-get upgrade -y
 
-echo "Reconfigure fonts."
+echo -e "${YEL}Reconfigure fonts.${NC}"
 sudo apt-get install -y ttf-dejavu ttf-liberation ttf-mscorefonts-installer xfonts-terminus
 sudo dpkg-reconfigure fontconfig-config
 sudo dpkg-reconfigure fontconfig
@@ -51,39 +55,50 @@ sudo dpkg-reconfigure fontconfig
   #touch ~/src/infinality/installed
 #fi
 
-echo "Installing Bumblebee Hybrid graphics."
+echo -e "${YEL}Installing Bumblebee Hybrid graphics.${NC}"
 sudo apt-get install -y bumblebee-nvidia primus mesa-utils
 sudo adduser $USER bumblebee
 #optirun glxgears -info
 
-echo "Installing RedShift blue-light management."
+echo -e "${YEL}Installing RedShift blue-light management.${NC}"
 sudo apt-get install -y redshift-plasmoid
 
-echo "Installing utils."
+echo -e "${YEL}Installing utils.${NC}"
 sudo apt-get install -y apg xclip
 
-echo "Installing common software."
+echo -e "${YEL}Installing common software.${NC}"
 sudo apt-get install -y plasma-widget-ktorrent
 
-if [ ! -f ~/src/installed ]
+if [ ! -f ~/src/installed ];
 then
-  echo "Downloading stand-alone packages."
   mkdir -p ~/src/
-  wget -O ~/src/atom-amd64.deb https://github.com/atom/atom/releases/download/v1.9.9/atom-amd64.deb
-  wget -O ~/src/google-chrome-stable_current_amd64.deb https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb
-  wget -O ~/src/steam_latest.deb https://steamcdn-a.akamaihd.net/client/installer/steam.deb
+  if [ ! -f ~/src/atom-amd64.deb ];
+  then
+    echo -e "${YEL}Downloading Atom.${NC}"
+    wget -O ~/src/atom-amd64.deb https://github.com/atom/atom/releases/download/v1.9.9/atom-amd64.deb
+  fi
+  if [ ! -f ~/src/google-chrome-stable_current_amd64.deb ];
+  then
+    echo -e "${YEL}Downloading Chrome.${NC}"
+    wget -O ~/src/google-chrome-stable_current_amd64.deb https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb
+  fi
+  if [ ! -f ~/src/steam_latest.deb ];
+  then
+    echo -e "${YEL}Downloading Steam.${NC}"
+    wget -O ~/src/steam_latest.deb https://steamcdn-a.akamaihd.net/client/installer/steam.deb
+  fi
 
-  echo "Installing stand-alone packages."
-  dpkg -i ~/src/atom-amd64.deb
-  dpkg -i ~/src/google-chrome-stable_current_amd64.deb
-  dpkg -i ~/src/steam_latest.deb
+  echo -e "${YEL}Installing stand-alone packages.${NC}"
+  sudo dpkg -i ~/src/atom-amd64.deb
+  sudo dpkg -i ~/src/google-chrome-stable_current_amd64.deb
+  sudo dpkg -i ~/src/steam_latest.deb
   sudo apt-get -fy install
   touch ~/src/installed
 fi
 
 if ! grep "GIT_PROMPT_ONLY_IN_REPO" ~/.bashrc;
 then
-  echo "Install GIT prompt."
+  echo -e "${YEL}Install GIT prompt.${NC}"
   echo "" >> ~/.bashrc
   echo "GIT_PROMPT_ONLY_IN_REPO=1" >> ~/.bashrc
   echo "source ~/scripts/git-prompt.sh" >> ~/.bashrc
